Paramerixed the sql-query, code refactring  for proper error handleing  in C#

+++ WebApplication/App_Code/studentCls.cs
@@ -14,7 +14,7 @@ namespace WebApplication.App_Code
         public string EducationalLvl { get; set; }
         public string Skill { get; set; }
         public string Email { get; set; }
-        public double ContactNo { get; set; }
+        public decimal ContactNo { get; set; }
         public int QuizeId { get; set; }
     }
 }
